+++ src/main/kotlin/config/Config.kt
@@ -2,7 +2,7 @@ package config
 
 import kotlinx.coroutines.flow.MutableStateFlow
 import lightOrgan.gateway.GatewayConfig
-import lightOrgan.spectralAnalyzer.spectrum.SpectralAnalyzerConfig
+import lightOrgan.spectralAnalysis.spectrum.SpectralAnalyzerConfig
 
 class Config(
     val startAutomatically: MutableStateFlow<Boolean>,
