+++ src/main/kotlin/gui/dashboard/DashboardViewModel.kt
@@ -4,14 +4,14 @@ import config.Config
 import config.ConfigPersister
 import config.ConfigSingleton
 import gui.dashboard.tiles.color.ColorTileViewModel
-import gui.dashboard.tiles.spectralAnalyzer.SpectralAnalysisTileViewModel
+import gui.dashboard.tiles.spectralAnalysis.SpectralAnalysisTileViewModel
 import gui.snackbar.SnackbarController
 import gui.tiles.audioInput.AudioInputTileViewModel
 import gui.tiles.gateway.GatewayTileViewModel
 import lightOrgan.color.ColorManager
 import lightOrgan.gateway.GatewayManager
 import lightOrgan.input.AudioInputManager
-import lightOrgan.spectralAnalyzer.SpectralAnalyzer
+import lightOrgan.spectralAnalysis.SpectralAnalyzer
 
 class DashboardViewModel(
     inputManager: AudioInputManager,
