+++ src/main/kotlin/lightOrgan/color/ColorManager.kt
@@ -5,7 +5,7 @@ import color.StandardRgbColors
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import lightOrgan.spectralAnalyzer.SpectralAnalysis
+import lightOrgan.spectralAnalysis.SpectralAnalysis
 
 // ENHANCEMENT: OKLCH (or other perceptually balanced spectrum)
 // ENHANCEMENT: Force a given hue, saturation, or color.
