+++ src/main/kotlin/lightOrgan/spectralAnalysis/peaks/PeakExtractor.kt
@@ -1,10 +1,10 @@
-package lightOrgan.spectralAnalyzer.peaks
+package lightOrgan.spectralAnalysis.peaks
 
 import dsp.bins.FrequencyBins
 import dsp.peakExtraction.ParabolicSpectralPeakExtractor
 import dsp.peakExtraction.SpectralPeakExtractor
 import dsp.peakExtraction.SpectralPeaks
-import lightOrgan.spectralAnalyzer.spectrum.SpectralAnalyzerConfig
+import lightOrgan.spectralAnalysis.spectrum.SpectralAnalyzerConfig
 
 // TODO: Test me
 // ENHANCEMENT: Reject peaks below the sidelobe dB? E.g. a peak of 0.5 yields sidelobes of X - then anything at X and below is removed
