+++ src/main/kotlin/lightOrgan/spectralAnalysis/spectrum/SpectralAnalyzerConfig.kt
@@ -1,4 +1,4 @@
-package lightOrgan.spectralAnalyzer.spectrum
+package lightOrgan.spectralAnalysis.spectrum
 
 import dsp.filtering.FilterConfig
 import dsp.windowing.WindowType
