+++ src/test/kotlin/lightOrgan/spectralAnalysis/FilterManagerTests.kt
@@ -1,4 +1,4 @@
-package lightOrgan.spectralAnalyzer
+package lightOrgan.spectralAnalysis
 
 import audio.samples.AudioFrame
 import dsp.filtering.Filter
@@ -8,7 +8,7 @@ import io.mockk.clearAllMocks
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
-import lightOrgan.spectralAnalyzer.spectrum.FilterManager
+import lightOrgan.spectralAnalysis.spectrum.FilterManager
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
